% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_gini}
\alias{calc_gini}
\title{Calculate the Gini coefficient (2 * AUC - 1)}
\usage{
calc_gini(actual, predicted_prob)
}
\arguments{
\item{actual}{Numeric 0/1 vector of true class labels.}

\item{predicted_prob}{Numeric vector of predicted class-1 probabilities.}
}
\value{
Single numeric in \verb{[-1, 1]}.
}
\description{
Calculate the Gini coefficient (2 * AUC - 1)
}
\keyword{internal}
