% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundary.R
\name{.is_closed}
\alias{.is_closed}
\title{Detect whether a polyline is closed (first row equals last row)}
\usage{
.is_closed(Mxy)
}
\arguments{
\item{Mxy}{Numeric matrix with at least 2 rows and 2 columns.}
}
\value{
Logical scalar.
}
\description{
Detect whether a polyline is closed (first row equals last row)
}
\keyword{internal}
