% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hull_utils.R
\name{.points_in_polygon}
\alias{.points_in_polygon}
\title{Test whether a set of points lies inside a SpatialPolygons hull}
\usage{
.points_in_polygon(Z, poly)
}
\arguments{
\item{Z}{n x 2 matrix of (x, y) coordinates.}

\item{poly}{An \code{sp::SpatialPolygons} object.}
}
\value{
Logical vector of length n.
}
\description{
Test whether a set of points lies inside a SpatialPolygons hull
}
\keyword{internal}
