% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_cf.R
\name{print.bl_filters}
\alias{print.bl_filters}
\title{Print method for bl_filters}
\usage{
\method{print}{bl_filters}(x, ...)
}
\arguments{
\item{x}{A \code{"bl_filters"} object.}

\item{...}{Currently ignored.}
}
\description{
Print method for bl_filters
}
