% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_cf.R
\name{print.bl_local_result}
\alias{print.bl_local_result}
\title{Print method for bl_local_result}
\usage{
\method{print}{bl_local_result}(x, ...)
}
\arguments{
\item{x}{A \code{"bl_local_result"} object.}

\item{...}{Currently ignored.}
}
\description{
Print method for bl_local_result
}
