% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapley.R
\name{print.bl_shapley}
\alias{print.bl_shapley}
\title{Print method for bl_shapley}
\usage{
\method{print}{bl_shapley}(x, ...)
}
\arguments{
\item{x}{A \code{"bl_shapley"} object.}

\item{...}{Currently ignored.}
}
\description{
Print method for bl_shapley
}
