% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapley.R
\name{print.bl_sparse_result}
\alias{print.bl_sparse_result}
\title{Print method for bl_sparse_result}
\usage{
\method{print}{bl_sparse_result}(x, ...)
}
\arguments{
\item{x}{A \code{"bl_sparse_result"} object.}

\item{...}{Currently ignored.}
}
\description{
Print method for bl_sparse_result
}
