% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_cf.R
\name{print.bl_target}
\alias{print.bl_target}
\title{Print method for bl_target}
\usage{
\method{print}{bl_target}(x, ...)
}
\arguments{
\item{x}{A \code{"bl_target"} object.}

\item{...}{Currently ignored.}
}
\description{
Print method for bl_target
}
