% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feasibility_utils.R
\name{get_filter_logical_vector}
\alias{get_filter_logical_vector}
\title{Evaluate a list of range expressions as a row-wise AND filter}
\usage{
get_filter_logical_vector(data, filter_exprs)
}
\arguments{
\item{data}{Data frame to evaluate against.}

\item{filter_exprs}{List of expressions as returned by
\code{get_variable_ranges()}.}
}
\value{
Logical vector of length \code{nrow(data)}. \code{TRUE} means the row
passes all filters.
}
\description{
Evaluate a list of range expressions as a row-wise AND filter
}
\keyword{internal}
