% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feasibility_utils.R
\name{get_variable_ranges}
\alias{get_variable_ranges}
\title{Get the min/max range of each variable in a data frame}
\usage{
get_variable_ranges(data)
}
\arguments{
\item{data}{A data frame. All columns are included.}
}
\value{
A list of expressions, one per column of \code{data}.
}
\description{
Returns a list of \code{rlang} expressions of the form
\code{var >= min & var <= max} for each column. These can be passed
directly to \code{get_filter_logical_vector()}.
}
\keyword{internal}
